+++ ui/apps/platform/src/Containers/Network/Wizard/Creator/Buttons/Generate.js
@@ -1,39 +1,80 @@
-import React, { Component } from 'react';
+import React from 'react';
 import PropTypes from 'prop-types';
+import { createStructuredSelector } from 'reselect';
 import { connect } from 'react-redux';
-import { actions as wizardActions } from 'reducers/network/wizard';
 
+import { selectors } from 'reducers';
+import { actions as wizardActions } from 'reducers/network/wizard';
+import FeatureEnabled from 'Containers/FeatureEnabled';
+import { knownBackendFlags } from 'utils/featureFlags';
 import wizardStages from '../../wizardStages';
 
-class GenerateButton extends Component {
-    static propTypes = {
-        setWizardStage: PropTypes.func.isRequired,
-        requestNetworkPolicyModification: PropTypes.func.isRequired,
-    };
+const GenerateButton = ({
+    setWizardStage,
+    requestNetworkPolicyModification,
+    excludePortsProtocols,
+    setExcludePortsProtocolsState,
+}) => {
+    function onClick() {
+        requestNetworkPolicyModification();
+        setWizardStage(wizardStages.simulator);
+    }
 
-    onClick = () => {
-        this.props.requestNetworkPolicyModification();
-        this.props.setWizardStage(wizardStages.simulator);
-    };
+    function onChangeHandler() {
+        setExcludePortsProtocolsState(!excludePortsProtocols);
+    }
 
-    render() {
-        return (
+    return (
+        <>
+            <FeatureEnabled featureFlag={knownBackendFlags.ROX_NETWORK_GRAPH_PORTS}>
+                {({ featureEnabled }) => {
+                    return (
+                        featureEnabled && (
+                            <div className="flex justify-center">
+                                <input
+                                    type="checkbox"
+                                    data-testid="checkbox-exclude-ports-protocols"
+                                    id="exclude-ports-protocols"
+                                    checked={!!excludePortsProtocols}
+                                    onChange={onChangeHandler}
+                                    aria-label="Exclude ports and protocols"
+                                />
+                                <label htmlFor="exclude-ports-protocols" className="pl-2">
+                                    Exclude ports & protocols
+                                </label>
+                            </div>
+                        )
+                    );
+                }}
+            </FeatureEnabled>
             <div className="flex m-3 py-2 items-center justify-center">
                 <button
                     type="button"
                     className="rounded-sm px-4 py-3 border-2 border-primary-300 hover:border-primary-400 text-center text-3xlg font-700 text-primary-700 bg-primary-100 hover:bg-primary-200"
-                    onClick={this.onClick}
+                    onClick={onClick}
                 >
                     Generate and simulate network policies
                 </button>
             </div>
-        );
-    }
-}
+        </>
+    );
+};
+
+GenerateButton.propTypes = {
+    setWizardStage: PropTypes.func.isRequired,
+    requestNetworkPolicyModification: PropTypes.func.isRequired,
+    excludePortsProtocols: PropTypes.bool.isRequired,
+    setExcludePortsProtocolsState: PropTypes.func.isRequired,
+};
+
+const mapStateToProps = createStructuredSelector({
+    excludePortsProtocols: selectors.getNetworkPolicyExcludePortsProtocolsState,
+});
 
 const mapDispatchToProps = {
     setWizardStage: wizardActions.setNetworkWizardStage,
     requestNetworkPolicyModification: wizardActions.generateNetworkPolicyModification,
+    setExcludePortsProtocolsState: wizardActions.setNetworkPolicyExcludePortsProtocolsState,
 };
 
-export default connect(null, mapDispatchToProps)(GenerateButton);
+export default connect(mapStateToProps, mapDispatchToProps)(GenerateButton);